requirements.txt
flask
firebase_admin
gunicorn
python-dotenv
requests
youtube-transcript-api
openai
paypalrestsdk
flask-cors
python-dateutil
setuptools
flask-wtf